Remove opensearch error taxonomy upload
The error taxonomy has been moved to the analytics DB, so we no longer
need to upload it to opensearch. Hopefully this also has the secondary
effect of lowering the stress on the opensearch cluster and causing
job log upload failures.
