+++ analytics/analytics/core/job_log_uploader/__init__.py
@@ -1,19 +1,19 @@
-import json
-import re
 from dataclasses import asdict
 from datetime import datetime
+import json
+import re
 from typing import Any
 
-import gitlab
 from celery import shared_task
 from django.conf import settings
+import gitlab
 from opensearch_dsl import Date, Document, connections
 from opensearchpy import ConnectionTimeout
 from requests.exceptions import ReadTimeout
 from urllib3.exceptions import ReadTimeoutError
 
 from analytics import setup_gitlab_job_sentry_tags
-from analytics.core.job_failure_classifier import _job_retry_data
+from analytics.job_processor.dimensions import _job_retry_data
 
 
 class JobLog(Document):
